- Lectura y escritura de ficheros vectoriales - Test

+++ src/geometry/entities/lineString.h
@@ -86,7 +86,7 @@ public:
    * \brief Longitud de la polilínea
    * \return Longitud
    */
-  size_t length() const;
+  double length() const;
 
   /*!
    * \brief Operador de asignación
@@ -136,9 +136,9 @@ LineString<Point_t>::LineString(std::initializer_list<Point_t> listPoints)
 //}
 
 template<typename Point_t> inline
-size_t LineString<Point_t>::length()  const
+double LineString<Point_t>::length()  const
 {
-  size_t length = 0.;
+  double length = 0.;
   for (size_t i = 1; i < this->mEntities.size(); i++) {
     length += distance(this->mEntities[i - 1], this->mEntities[i]);
   }
